Added a base parameter
The base parameter should make it easier to normalize the path. Sometimes NFS resources are mounted using an alias, and other times they are mounted using the full path to the export. Standardizing on what is normally seen as the alias, makes it easier to group the various systems that show up under multiple groups otherwise.